+++ src/events/schemas/joi.create-event.schema.ts
@@ -4,6 +4,7 @@ import { joiObjectId } from '@/common/schemas/joi.object-id.schema';
 import { joiPriceGroupSchema } from '@/common/schemas/joi.price-group.schema';
 import { EventStatus } from '../enums/event-status.enum';
 import { EventType } from '../enums/event-type.enum';
+import { joiSlugSchema } from '@/common/schemas/joi.slug.schema';
 
 export const createEventSchema = Joi.object({
   domainId: joiObjectId.required(),
@@ -19,13 +20,7 @@ export const createEventSchema = Joi.object({
 
   title: Joi.string().trim().min(1).max(150).required(),
 
-  slug: Joi.string()
-    .trim()
-    .lowercase()
-    .pattern(/^[a-z0-9-]+$/)
-    .min(1)
-    .max(120)
-    .required(),
+  slug: joiSlugSchema.required(),
 
   startAt: joiUtcIsoDateTime.required(),
   endAt: joiUtcIsoDateTime.required(),