+++ src/menus/menus.module.ts
@@ -2,9 +2,9 @@ import { Module } from '@nestjs/common';
 import { MongooseModule } from '@nestjs/mongoose';
 import { DomainsModule } from '@/domains/domains.module';
 import { PagesModule } from '@/pages/pages.module';
-import { Menu, menuSchema } from './schemas/menu.schema';
 import { MenusController } from './menus.controller';
 import { MenusService } from './menus.service';
+import { Menu, menuSchema } from './schemas/menu.schema';
 
 @Module({
   imports: [